load uinput kernel module
Without the uinput module, "uaccess" udev tag is not applied to
`/dev/uinput`, which means neither Steam Controller nor any gamepad
handled by Steam Input is working (only in Steam menus, but not in
game). Steam complains:

> Couldn't initialize virtual gamepad: Couldn't open /dev/uinput for
writing

Without this fix, the user has to be in the `input` group. With
this fix, it's no longer necessary and everything is working out of the
box.